[Image-Example] refactor: add styled Image to example app
